Split Bare into multiple modules, refactor
