+++ src/Language/Haskell/Liquid/Errors.hs
@@ -228,6 +228,11 @@ ppError' _ dSp (ErrMismatch _ x τ t)
     $+$ text "Haskell:" <+> pprint τ
     $+$ text "Liquid :" <+> pprint t
 
+ppError' _ dSp (ErrIllegalAliasApp _ dn dl)
+  = dSp <+> text "Refinement Type Alias cannot be used in this context"
+    $+$ text "Type alias:" <+> pprint dn
+    $+$ text "Defined at:" <+> pprint dl
+
 ppError' _ dSp (ErrAliasApp _ n name dl dn)
   = dSp <+> text "Malformed Type Alias Application"
     $+$ text "Type alias:" <+> pprint name
